Broke out the appController into three different controllers, appointments, users, and packages. Also cleaned up and alphabetized the API file, and created nested folders within the component folder for ClientSide and PhotographerSide to make it a little easier to develop.
