+++ client/src/components/ClientSide/EnterAddress/EnterAddress.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { useHistory } from 'react-router-dom';
-import API from '../../utils/API';
+import API from '../../../utils/API';
 import './EnterAddress.css';
 import { CurrentStep } from '../CurrentStep/CurrentStep';
-import getUserId from '../../utils/getUserId';
+import getUserId from '../../../utils/getUserId';
 
 export function EnterAddress({ appointmentData, setAppointmentData }) {
