+++ client/src/components/PhotographerSide/LeftNav/LeftNav.js
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import UserContext from "../../utils/UserContext";
+import UserContext from "../../../utils/UserContext";
 import { useHistory } from 'react-router-dom';
 import './LeftNav.css';
