+++ client/src/utils/API.js
@@ -1,11 +1,21 @@
 import axios from "axios";
 
 export default {
+  cancelEvent: function (id) {
+    return axios.post(`/api/appointments/cancel/:${id}`, {
+      params: {
+        _id: id
+      }
+    })
+  },
+  createNewPackage: function (id, newPackage) {
+    return axios.post("/api/packages/new", newPackage);
+  },
+  createNewUser: function (newUser) {
+    return axios.post("/api/users", newUser);
+  },
   getAddress: function (options) {
     return axios.request(options)
-  },
-  saveAppointment: function (appointmentData) {
-    return axios.post("/api/appointments", appointmentData);
   },
   getAppointments: async function (appointmentData) {
     return await axios.get("/api/appointments", {
@@ -14,31 +24,10 @@ export default {
       }
     });
   },
-  createNewUser: function (newUser) {
-    return axios.post("/api/users", newUser);
-  },
-  createNewPackage: function (id, newPackage) {
-    return axios.post("/api/packages/new", newPackage);
-  },
-  loginUser: function (userLoginData) {
-    return axios.get("/api/users/", {
-      params: {
-        email: userLoginData.email,
-        password: userLoginData.password
-      }
-    });
-  },
   getDashboardData: function (userId) {
     return axios.get('/api/appointments/dashboard', {
       params: {
         photographerId: userId
-      }
-    })
-  },
-  getUserData: function (userId) {
-    return axios.get('/api/users/dashboard', {
-      params: {
-        _id: userId
       }
     })
   },
@@ -50,19 +39,10 @@ export default {
       }
     })
   },
-  cancelEvent: function (id) {
-    return axios.post(`/api/appointments/cancel/:${id}`, {
+  getUserData: function (userId) {
+    return axios.get('/api/users/dashboard', {
       params: {
-        _id: id
-      }
-    })
-  },
-  updateUser: function (id, update) {
-    console.log(update)
-    return axios.post(`api/users/update/:${id}`, {
-      params: {
-        _id: id,
-        update
+        _id: userId
       }
     })
   },
@@ -70,6 +50,26 @@ export default {
     return axios.get(`/api/users/${userEmail}`, {
       params: {
         email: userEmail
+      }
+    })
+  },
+  loginUser: function (userLoginData) {
+    return axios.get("/api/users/", {
+      params: {
+        email: userLoginData.email,
+        password: userLoginData.password
+      }
+    });
+  },
+  saveAppointment: function (appointmentData) {
+    return axios.post("/api/appointments", appointmentData);
+  },
+  updateUser: function (id, update) {
+    console.log(update)
+    return axios.post(`api/users/update/:${id}`, {
+      params: {
+        _id: id,
+        update
       }
     })
   },
