--- client/src/utils/getUserFromLocalStorage
@@ -1,10 +1,0 @@
-
-export default {
-    getUserFromLocalStorage: function () {
-
-        const userId = pathname.substring(pathname.lastIndexOf('/') + 1);
-
-        return userId;
-    },
-
-}
